requirements.txt
fastapi
uvicorn
langchain
openai
sqlalchemy
pyodbc
pandas
python-dotenv
